Multiple changes to Graph and tests
Changed the internal nodes list in Graph, making it private and
renaming it to _nodes, while adding a public property Nodes that
returns a ReadOnlyCollection. Also removed the GetNodes() method as it
is replaced by a property.

Added a Count property, together with a public readonly struct
GraphCount that contains a field for the number of nodes and one for
the number of connections.

Marked NodeNotFoundException as public.

Changed a bunch of the tests to instead use the new Nodes property.

Added a bunch of unit tests for RemoveNode.