Got monogame FINALLY working
To get it too work i a new project in the same solution named NodeMonog. This project was prebuilt with vs2017 monogame preset so everything just works. I can also merge it well since it is in a seperate project. I was tunnelled vision into making it a new cs file in the main project, which would have made sense in some ways, but man is this better (since it works and all). So now I can finally start designing and creatign a fitting GUI.