Added the Force-based graph layout, +methods to Graph and Connection
Added Physics.cs together with Core.Physics, which handles simulating
the physics system for use in making a layout.

Added a property Connections to Graph for simply enumerating all the
existing connections.

Added methods to Graph, namely:
- (private) FindNodeIndex
- TryGetDirectedConnection
- GetNeighbors

Added a virtual Strength method to Connection, to give a numerical
indicator of the overall strength the Connection has.