Various small changes
Further improved startup UI and simplified code

Traits/Statuses now start as null in Connections

Added an ugly hack to family stuff so that it doesn't try to access
the second member of a 1-person family.

Changed a confusing error/status message when simulations get
cancelled.

Fixed a minor bug in Renderer